requirements.txt
streamlit
opencv-python
face-recognition
pandas
numpy
Pillow
